+++ app/src/main/java/xyz/narengi/android/ui/fragment/HouseGuestEntryFragment.java
@@ -229,7 +229,7 @@ public class HouseGuestEntryFragment extends HouseEntryBaseFragment {
         if (priceText == null || priceText.length() == 0 || getHouse() == null)
             return;
 
-        Double price = Double.parseDouble(priceText);
+        int price = Integer.parseInt(priceText);
         if (getHouse().getPrice() == null) {
             HousePrice housePrice = new HousePrice();
             getHouse().setPrice(housePrice);
